+++ forge/src/main/java/net/mehvahdjukaar/polytone/forge/PolytoneForge.java
@@ -1,5 +1,6 @@
 package net.mehvahdjukaar.polytone.forge;
 
+import com.github.alexmodguy.alexscaves.mixin.client.LightTextureMixin;
 import com.google.common.collect.Lists;
 import com.mojang.blaze3d.vertex.PoseStack;
 import net.mehvahdjukaar.polytone.Polytone;
@@ -8,6 +9,7 @@ import net.mehvahdjukaar.polytone.slotify.SlotifyScreen;
 import net.mehvahdjukaar.polytone.tabs.ItemToTabEvent;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.screens.Screen;
+import net.minecraft.client.renderer.LightTexture;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.world.item.CreativeModeTab;
 import net.minecraft.world.item.ItemStack;