+++ changelog.md
@@ -1,1 +1,6 @@
- fabric render types now accept cutout_mipped and tripwire, making up all the ones that can be found in chunk buffer
+- added alex caves compat for lightmaps
+- added a dedicated polytone log file at /logs/polytone.log
+- changed resource reload toast when its caused by polytone
+- improved logging for colormaps missing textures
+- fabric now has support for fog fluid colors
+- converted optifine fluid properties will be applied to flowing fluids aswell on fabric